Exit immediately if --parent-pid doesn't look like an integer.
If you pass "--parent-pid foobar", then we immediately log an error
message and exit. This check happens via `parseInt` rather than testing
for exceptions in `process.kill(parentPid, 0)` because `process.kill`
converts its pid argument to an integer, so `process.kill('foobar', 0)`
returns true just as `process.kill(0, 0)` does.

+++ packages/webapp/webapp_server.js
@@ -69,11 +69,17 @@ var initKeepalive = function () {
 //   takes the parent process's place before the child process dies.
 var startCheckForLiveParent = function (parentPid) {
   if (parentPid) {
+    // Check that we have a pid that looks like an integer.
+    if (+parseInt(parentPid) !== parentPid) {
+      console.error("--parent-pid must be a valid process ID.");
+      process.exit(1);
+    }
+
     setInterval(function () {
       try {
         process.kill(parentPid, 0);
       } catch (err) {
-        console.log("Parent process is dead! Exiting.");
+        console.error("Parent process is dead! Exiting.");
         process.exit(1);
       }
     });
